refactor: fix after merge & rename packages
