+++ metals/src/main/scala/scala/meta/internal/metals/doctor/Doctor.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.metals
+package scala.meta.internal.metals.doctor
 
 import java.net.URLEncoder
 import java.nio.charset.StandardCharsets
@@ -13,11 +13,28 @@ import scala.meta.internal.bsp.ResolvedBloop
 import scala.meta.internal.bsp.ResolvedBspOne
 import scala.meta.internal.bsp.ResolvedMultiple
 import scala.meta.internal.bsp.ResolvedNone
+import scala.meta.internal.metals.BuildInfo
+import scala.meta.internal.metals.BuildTargets
+import scala.meta.internal.metals.ClientCommands
+import scala.meta.internal.metals.ClientConfiguration
+import scala.meta.internal.metals.Diagnostics
+import scala.meta.internal.metals.FileDecoderProvider
+import scala.meta.internal.metals.HtmlBuilder
+import scala.meta.internal.metals.Icons
+import scala.meta.internal.metals.JavaTarget
+import scala.meta.internal.metals.JdkSources
+import scala.meta.internal.metals.JdkVersion
 import scala.meta.internal.metals.Messages.CheckDoctor
 import scala.meta.internal.metals.MetalsEnrichments._
+import scala.meta.internal.metals.MetalsHttpServer
+import scala.meta.internal.metals.MtagsResolver
+import scala.meta.internal.metals.ParametrizedCommand
+import scala.meta.internal.metals.PopupChoiceReset
+import scala.meta.internal.metals.ScalaTarget
+import scala.meta.internal.metals.Tables
+import scala.meta.internal.metals.Urls
 import scala.meta.internal.metals.clients.language.MetalsLanguageClient
 import scala.meta.internal.metals.config.DoctorFormat
-import scala.meta.internal.troubleshoot.ProblemResolver
 import scala.meta.io.AbsolutePath
 
 import ch.epfl.scala.bsp4j.BuildTargetIdentifier
@@ -249,15 +266,18 @@ final class Doctor(
 
     val (buildServerHeading, _) = selectedBuildServerMessage()
     val importBuildHeading = selectedImportBuildMessage()
+    val jdkInfo = getJdkInfo()
+    val serverInfo = s"Metals server version: ${BuildInfo.metalsVersion}"
     val heading =
       List(
         buildToolHeading,
         Some(buildServerHeading),
         importBuildHeading,
+        jdkInfo,
+        Some(serverInfo),
         Some(doctorHeading)
       ).flatten
         .mkString("\n\n")
-
     val results = if (targetIds.isEmpty) {
       DoctorResults(
         doctorTitle,
