+++ metals/src/main/scala/scala/meta/internal/metals/doctor/DoctorResults.scala
@@ -1,4 +1,6 @@
-package scala.meta.internal.metals
+package scala.meta.internal.metals.doctor
+
+import scala.meta.internal.metals.Icons
 
 import ujson.Obj
