+++ metals/src/main/scala/scala/meta/internal/metals/doctor/JavaProblem.scala
@@ -1,4 +1,4 @@
-package scala.meta.internal.troubleshoot
+package scala.meta.internal.metals.doctor
 
 import scala.meta.internal.metals.BuildInfo
