+++ tests/unit/src/test/scala/tests/troubleshoot/ProblemResolverSuite.scala
@@ -7,18 +7,18 @@ import scala.meta.internal.metals.BuildInfo
 import scala.meta.internal.metals.MetalsEnrichments._
 import scala.meta.internal.metals.ScalaTarget
 import scala.meta.internal.metals.ScalaVersions
-import scala.meta.internal.troubleshoot.DeprecatedSbtVersion
-import scala.meta.internal.troubleshoot.DeprecatedScalaVersion
-import scala.meta.internal.troubleshoot.FutureSbtVersion
-import scala.meta.internal.troubleshoot.FutureScalaVersion
-import scala.meta.internal.troubleshoot.MissingJdkSources
-import scala.meta.internal.troubleshoot.MissingSourceRoot
-import scala.meta.internal.troubleshoot.OutdatedJunitInterfaceVersion
-import scala.meta.internal.troubleshoot.OutdatedMunitInterfaceVersion
-import scala.meta.internal.troubleshoot.ProblemResolver
-import scala.meta.internal.troubleshoot.SemanticDBDisabled
-import scala.meta.internal.troubleshoot.UnsupportedSbtVersion
-import scala.meta.internal.troubleshoot.UnsupportedScalaVersion
+import scala.meta.internal.metals.doctor.DeprecatedSbtVersion
+import scala.meta.internal.metals.doctor.DeprecatedScalaVersion
+import scala.meta.internal.metals.doctor.FutureSbtVersion
+import scala.meta.internal.metals.doctor.FutureScalaVersion
+import scala.meta.internal.metals.doctor.MissingJdkSources
+import scala.meta.internal.metals.doctor.MissingSourceRoot
+import scala.meta.internal.metals.doctor.OutdatedJunitInterfaceVersion
+import scala.meta.internal.metals.doctor.OutdatedMunitInterfaceVersion
+import scala.meta.internal.metals.doctor.ProblemResolver
+import scala.meta.internal.metals.doctor.SemanticDBDisabled
+import scala.meta.internal.metals.doctor.UnsupportedSbtVersion
+import scala.meta.internal.metals.doctor.UnsupportedScalaVersion
 import scala.meta.io.AbsolutePath
 
 import ch.epfl.scala.bsp4j.BuildTarget
